requirements.txt
streamlit>=1.33.0
boto3>=1.34.72
requests>=2.32.3
beautifulsoup4==4.12.2
python-dotenv==1.0.0
PyPDF2==3.0.1
protobuf>=5.0,<6.0
tenacity>=9.0.0,<10.0.0
